+++ src/Support/Adapters/ExcelExport.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Sfneal\ViewExport\Support\Adapters;
+
+abstract class ExcelExport
+{
+}