+++ src/Support/Adapters/ExportService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sfneal\ViewExport\Support;
+namespace Sfneal\ViewExport\Support\Adapters;
 
 use Illuminate\Contracts\View\View;
 use Sfneal\ViewModels\AbstractViewModel;
@@ -8,7 +8,7 @@ use Sfneal\ViewModels\AbstractViewModel;
 abstract class ExportService
 {
     /**
-     * Provide a view to build the PDF from.
+     * Provide a view to build the export from.
      *
      * @param View $view
      * @param string|null $uploadPath
@@ -17,7 +17,7 @@ abstract class ExportService
     abstract public static function fromView(View $view, string $uploadPath = null): Renderer;
 
     /**
-     * Provide a view to build the PDF from.
+     * Provide a view to build the export from.
      *
      * @param AbstractViewModel $viewModel
      * @param string|null $uploadPath