+++ src/Support/Adapters/Renderer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sfneal\ViewExport\Support;
+namespace Sfneal\ViewExport\Support\Adapters;
 
 use Illuminate\Support\Facades\Bus;
 use Sfneal\Queueables\AbstractJob;
@@ -13,7 +13,7 @@ abstract class Renderer extends AbstractJob
     protected $content;
 
     /**
-     * @var string|null AWS S3 path to upload PDF to after render (if provided)
+     * @var string|null AWS S3 path to upload a file to after render (if provided)
      */
     protected $uploadPath;